requirements.txt
pandas
numpy
pymongo
flask
flask-cors
python-dotenv
faker